+++ docs/CHANGELOG.md
@@ -4,6 +4,23 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [5.2.0] - 2024-04-10
+
+### Added
+
+- Command line argument (-C) to configure and run Harbinger in single command
+- Commands to start the Minecraft Server
+- Command to send embed to contextual channel including server IP address and client version
+- Minecraft Server command STDOUT prints to ctx.channel when Minecraft command sent to server
+
+### Changed
+
+- Directory paths stored in config.ini now expand (~) characters to correct absolute paths
+
+### Fixed
+
+- Broken server directory path concatenation
+
 ## [5.1.0] - 2024-04-09
 
 ### Added